Start work on reducing file size - use 32bit macro time stamp and rollovers

+++ TcspcEvent.h
@@ -13,13 +13,13 @@ public:
       FrameMarker = 8
    };
 
-   uint64_t macro_time;
-   uint16_t micro_time;
+   uint32_t macro_time;
+   uint32_t micro_time;
    uint8_t channel;
    uint8_t mark;
 
 protected:
-
+  
    template<class T>
    T readBits(T& p, long n_bits)
    {
